getThreadName
